pyproject.toml
[project]
name = "domo-tech-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "fastapi[standard]>=0.121.2",
    "google-auth>=2.43.0",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-groq>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "markdown>=3.10",
    "motor>=3.7.1",
    "pymongo[srv]>=4.15.4",
    "pymupdf>=1.26.6",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sentence-transformers>=5.1.2",
    "unstructured>=0.18.20",
    "uvicorn>=0.38.0",
]
